requirements.txt
fastapi
uvicorn
python-dotenv
pyTelegramBotAPI